// Web/src/generated/model/deploymentPlanOwnerKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ReleaseHub API
 * ReleaseHub 前後端共用的 HTTP API 契約。
 * OpenAPI spec version: 1.0.0
 */

export type DeploymentPlanOwnerKind = typeof DeploymentPlanOwnerKind[keyof typeof DeploymentPlanOwnerKind];


export const DeploymentPlanOwnerKind = {
  platform: 'platform',
  project: 'project',
} as const;